refactor: separate environment to standalone package
