+++ README.md
@@ -1,14 +1,17 @@
-# RL Memory
+# grid-world-plus
 
-![Python 3.7+] ![Contributions welcome][contributions-badge] [![License: MIT][mit-license-badge]][license] ![PyTorch][torch-badge]
+![Python 3.8+] ![Contributions welcome][contributions-badge] [![License: MIT][mit-license-badge]][license] ![PyTorch][torch-badge]
 
 
 A research project in vision-based deep reinforcement learning centered on transfer learning, unsupervised representation learning, and applying attention mechanisms to interpret memories.
 
+- TODO describe pkg `grid_world_plus`
+- TODO describe pkg `rl-algos`
 
-[Python 3.7+]: https://img.shields.io/badge/python-3.7+-blue.svg
+
+[Python 3.8+]: https://img.shields.io/badge/python-3.7+-blue.svg
 [mit-license-badge]: https://img.shields.io/badge/License-MIT-g.svg 
-[license]: https://github.com/eskalnes/RL_memory/blob/main/LICENSE
+[license]: https://github.com/eskalnes/grid_world_plus/blob/main/LICENSE
 [contributions-badge]: https://img.shields.io/badge/contributions-welcome-yellow.svg 
 [torch-badge]: https://img.shields.io/badge/PyTorch-%23EE4C2C.svg?&logo=PyTorch&logoColor=white
 
@@ -17,8 +20,8 @@ A research project in vision-based deep reinforcement learning centered on trans
 - [[Poster]][poster-link] & [[poster presentation]][poster-presentation]: For a high-level overview.
 
 
-[report-link]: https://raw.githack.com/Unique-Divine/RL_memory/main/notes/presentations/reportS21-RL_Memory.pdf
-[poster-link]: https://raw.githack.com/Unique-Divine/RL_memory/main/notes/presentations/seminar_poster-rl_memory.pdf
+[report-link]: https://raw.githack.com/Unique-Divine/grid_world_plus/main/notes/presentations/reportS21-grid_world_plus.pdf
+[poster-link]: https://raw.githack.com/Unique-Divine/grid_world_plus/main/notes/presentations/seminar_poster-grid_world_plus.pdf
 [poster-presentation]: https://youtu.be/6iCXpgfTO6A
