+++ grid-world-plus/pyproject.toml
@@ -1,0 +1,26 @@
+[tool.poetry]
+name = "grid_world_plus"
+version = "0.0.0"
+description = "An environment consisting of a 2-dimenionsal grid of varying size"
+authors = ["Unique-Divine <realuniquedivine@gmail.com>"]
+license = "MIT"
+
+[tool.poetry.dependencies]
+python = "^3.9"
+numpy = "^1.23.1"
+matplotlib = "^3.5.2"
+
+[tool.poe.tasks]
+## PyTorch with CUDA 11.1. If PyTorch is imported first, 
+## importing Tensorflow will detect CUDA + cuDNN bundled with PyTorch
+## Run with the command "poe force-cuda11"
+## See https://github.com/python-poetry/poetry/issues/2543
+force-cuda11 = "pip install torch==1.9.0+cu111 torchvision==0.10.0+cu111 -f https://download.pytorch.org/whl/torch_stable.html"
+
+[tool.poetry.dev-dependencies]
+black = "^21.6b0"
+poethepoet = "^0.10.0"
+
+[build-system]
+requires = ["poetry-core>=1.0.0"]
+build-backend = "poetry.core.masonry.api"
