--- rl_memory/models/a2c/tools.py
@@ -1,12 +1,0 @@
-import numpy as np
-
-def discounted_reward(r, gamma):
-    """ take 1D float array of rewards and compute discounted reward """
-    discounted_r = np.zeros_like(r)
-    running_sum = 0
-    for i in reversed(range(0,len(r))):
-        discounted_r[i] = running_sum * gamma + r[i]
-        running_sum = discounted_r[i]
-    return list(discounted_r)
-
-
