--- rl_memory/models/archive/a2c copy/tools.py
@@ -1,42 +1,0 @@
-import numpy as np
-import matplotlib.pyplot as plt
-
-
-def discounted_reward(r, gamma):
-    """ take 1D float array of rewards and compute discounted reward """
-    discounted_r = np.zeros_like(r)
-    running_sum = 0
-    for i in reversed(range(0,len(r))):
-        discounted_r[i] = running_sum * gamma + r[i]
-        running_sum = discounted_r[i]
-    return list(discounted_r)
-
-
-def plot_episode_rewards(values, title='', show: bool = False):
-    """ Plot the reward curve and histogram of results over time."""
-
-    # Define the figure
-    f, ax = plt.subplots(nrows=1, ncols=2, figsize=(12,5))
-    f.suptitle(title)
-    ax[0].plot(values, label='score per run')
-    ax[0].axhline(195, c='red', ls='--', label='goal')
-    ax[0].set_xlabel('Episodes')
-    ax[0].set_ylabel('Reward')
-    x = range(len(values))
-    ax[0].legend()
-    # Calculate the trend
-    try:
-        z = np.polyfit(x, values, 1)
-        p = np.poly1d(z)
-        ax[0].plot(x, p(x), "--", label='trend')
-    except:
-        print('')
-
-    # Plot the histogram of results
-    ax[1].hist(values[-50:])
-    ax[1].axvline(195, c='red', label='goal')
-    ax[1].set_xlabel('Scores per Last 50 Episodes')
-    ax[1].set_ylabel('Frequency')
-    ax[1].legend()
-    if show:
-        plt.show()
